Create the UserService
At this point, I customised the insertUser method of the UserService because in the original implementation the author created in the service a new UUID for the user. I figure this doesn't make much sense since the User entity itself already provides this information, and it should not simply be disregarded and overridden by the service class.